[ui-core] move hooks to nested folder
