+++ desktop/core/src/desktop/js/apps/storageBrowser/StorageDirectoryPage/CreateAndUploadAction/CreateAndUploadAction.test.tsx
@@ -8,7 +8,7 @@ import {
 } from '../../../../reactComponents/FileChooser/api';
 
 const mockSave = jest.fn();
-jest.mock('../../../../utils/hooks/useSaveData', () => ({
+jest.mock('../../../../utils/hooks/useSaveData/useSaveData', () => ({
   __esModule: true,
   default: jest.fn(() => ({
     save: mockSave
