+++ desktop/core/src/desktop/js/reactComponents/GlobalAlert/GlobalAlert.tsx
@@ -26,7 +26,7 @@ import {
   HIDE_GLOBAL_ALERTS_TOPIC
 } from './events';
 import { HueAlert } from './types';
-import { useHuePubSub } from '../../utils/hooks/useHuePubSub';
+import { useHuePubSub } from '../../utils/hooks/useHuePubSub/useHuePubSub';
 import { i18nReact } from 'utils/i18nReact';
 
 type alertType = AlertProps['type'];
